+++ docs/pages/x/api/charts/charts-axis-highlight.json
@@ -3,7 +3,7 @@
   "slots": {},
   "name": "ChartsAxisHighlight",
   "styles": { "classes": [], "globalClasses": {}, "name": "MuiChartsAxisHighlight" },
-  "filename": "/packages/x-charts/src/AxisHighlight/ChartsAxisHighlight.tsx",
+  "filename": "/packages/x-charts/src/ChartsAxisHighlight/ChartsAxisHighlight.tsx",
   "demos": "<ul></ul>",
   "packages": [{ "packageName": "@mui/x-charts", "componentName": "ChartsAxisHighlight" }]
 }
