+++ docs/pages/x/api/charts/charts-clip-path.json
@@ -3,7 +3,7 @@
   "slots": {},
   "name": "ChartsClipPath",
   "styles": { "classes": [], "globalClasses": {}, "name": "MuiChartsClipPath" },
-  "filename": "/packages/x-charts/src/ClipPath/ChartsClipPath.tsx",
+  "filename": "/packages/x-charts/src/ChartsClipPath/ChartsClipPath.tsx",
   "demos": "<ul></ul>",
   "packages": [{ "packageName": "@mui/x-charts", "componentName": "ChartsClipPath" }]
 }
