+++ packages/x-charts/src/BarChart/BarChart.tsx
@@ -6,14 +6,14 @@ import {
   ResponsiveChartContainer,
   ResponsiveChartContainerProps,
 } from '../ResponsiveChartContainer';
-import { ChartsAxis, ChartsAxisProps } from '../Axis';
+import { ChartsAxis, ChartsAxisProps } from '../ChartsAxis';
 import { BarSeriesType } from '../models/seriesType/bar';
 import { MakeOptional } from '../models/helpers';
 import { DEFAULT_X_AXIS_KEY } from '../constants';
 import { ChartsTooltip, ChartsTooltipProps } from '../ChartsTooltip';
-import { ChartsLegend, ChartsLegendProps } from '../Legend';
-import { ChartsAxisHighlight, ChartsAxisHighlightProps } from '../AxisHighlight';
-import { ChartsClipPath } from '../ClipPath';
+import { ChartsLegend, ChartsLegendProps } from '../ChartsLegend';
+import { ChartsAxisHighlight, ChartsAxisHighlightProps } from '../ChartsAxisHighlight';
+import { ChartsClipPath } from '../ChartsClipPath';
 
 export interface BarChartProps
   extends Omit<ResponsiveChartContainerProps, 'series'>,
