+++ packages/x-charts/src/ChartsClipPath/ChartsClipPath.tsx
