+++ packages/x-charts/src/ChartsClipPath/index.tsx
