+++ packages/x-charts/src/ChartsLegend/utils.ts
