+++ packages/x-charts/src/index.ts
@@ -1,11 +1,11 @@
 export * from './constants';
 export * from './context';
-export * from './ClipPath';
-export * from './Axis';
-export * from './XAxis';
-export * from './YAxis';
+export * from './ChartsClipPath';
+export * from './ChartsAxis';
+export * from './ChartsXAxis';
+export * from './ChartsYAxis';
 export * from './ChartsTooltip';
-export * from './AxisHighlight';
+export * from './ChartsAxisHighlight';
 export * from './BarChart';
 export * from './LineChart';
 export * from './ScatterChart';
