+++ packages/x-charts/src/internals/components/AxisSharedComponents.tsx
@@ -1,5 +1,5 @@
 import { styled } from '@mui/material/styles';
-import { axisClasses } from '../../Axis/axisClasses';
+import { axisClasses } from '../../ChartsAxis/axisClasses';
 
 export const AxisRoot = styled('g', {
   name: 'MuiChartsAxis',
